requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
Deprecated==1.2.5
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
keras-metrics==1.0.0
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.2
pandas==0.24.2
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
PyGithub==1.43.5
PyJWT==1.7.1
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.15.1
wrapt==1.11.1
